Collectors: Raise warning when file glob is empty
When a file_glob in collect_files() does not match any file,
we raise a warning. This is so that users do not use unnecessary
globs and do not get confused why a glob doesnt work.
We modify the icollect() function to return the glob where the
file was found and handle the output appropriately in all places
it is called.
Fixes https://github.com/coala-analyzer/coala/issues/1520