Lint: Add helper to auto generate configs
This commit adds config_file() which allows LintBear creators
to make their own config files and return it as a list of lines.
This is then used in generate_config_file() to generate a temporary
file which is used as a config file. The temporary config file can be
used in arguments using {config_file}.

Fixes https://github.com/coala-analyzer/coala/issues/1475