MarkdownBear: Add options for remark
Add all the options that remark provides. Let Lint create
a config file and give that to remark.
The remark configs for cli are complicated to setup, and it
is much easier to use the JSON configuration file which remark
accepts. Hence let Lint take the configs that are required and
give those to remark.

Fixes https://github.com/coala-analyzer/coala/issues/1415